requirements.txt
pyyaml

# API
fastapi==0.75.0
Jinja2==3.0.3
requests==2.27.1
uvicorn==0.17.6

# Database
alembic==1.7.7
asyncpg==0.25.0
databases==0.5.5
psycopg2-binary==2.9.3
SQLAlchemy==1.4.32
pydantic

# async Unittest
#trio==0.20.0
# anyio==3.5.0
# aiosqlite
